+++ neuron/src/app/Neuron/Config.hs
@@ -18,29 +18,35 @@ module Neuron.Config
 where
 
 import Data.Either.Validation (validationToEither)
+import qualified Data.Text as T
 import Development.Shake (Action, readFile')
 import Dhall (FromDhall)
-import qualified Dhall
-import qualified Dhall.Core
-import qualified Dhall.Parser
-import qualified Dhall.TypeCheck
+import qualified Dhall (Decoder (extract), auto)
+import qualified Dhall.Core (normalize)
+import qualified Dhall.Parser (exprFromText)
+import qualified Dhall.TypeCheck (typeOf)
 import Neuron.Config.Orphans ()
 import Neuron.Config.Type (Config, configFile, defaultConfig, mergeWithDefault)
 import Relude
 import Rib.Shake (ribInputDir)
-import System.Directory
-import System.FilePath
+import System.Directory (doesFileExist)
+import System.FilePath ((</>))
 
 -- | Read the optional @neuron.dhall@ config file from the zettelkasten
 getConfig :: Action Config
 getConfig = do
-  configPath <- ribInputDir <&> (</> configFile)
+  notesDir <- ribInputDir
+  let configPath = notesDir </> configFile
   configVal :: Text <-
     liftIO (doesFileExist configPath) >>= \case
       True -> do
-        mergeWithDefault . toText <$> readFile' configPath
+        s <- readFile' configPath
+        -- Accept empty neuron.dhall (used to signify a directory to be used with neuron)
+        if T.null (T.strip $ toText s)
+          then pure defaultConfig
+          else mergeWithDefault . toText <$> readFile' configPath
       False ->
-        pure defaultConfig
+        fail $ "not a neuron notes directory (no neuron.dhall found under " <> notesDir <> ")"
   either fail pure $ parsePure configFile $ mergeWithDefault configVal
 
 -- | Pure version of `Dhall.input Dhall.auto`