+++ CHANGELOG.md
@@ -25,6 +25,7 @@
 - CLI
   - Faster querying: add `--cached` option to `neuron query`, to run faster using the cache. To keep the cache up to date, make sure that `neuron rib` is running.
   - Add `--id` and `--search` options to `open` command to open given zettel ID or search page respectively  (#317)
+  - Use current directory as the default value for `-d` (a la. git) (#389)
 
 ## 0.6.0.0